Continue cleanup. Removed SN module dependencies,  and old files
